backend/blog_api/src/main.py
import os
import sys
from dotenv import load_dotenv
from flask import Flask, jsonify, send_from_directory
from flask_cors import CORS

# Carregar variáveis de ambiente
load_dotenv()

# DON'T CHANGE THIS !!!
sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))

from src.models.user import db
from src.routes.user import user_bp
from src.routes.firebase_proxy import firebase_bp
from src.routes.contact import contact_bp
from src.routes.newsletter import newsletter_bp

app = Flask(__name__, static_folder=os.path.join(os.path.dirname(__file__), 'static'))
app.config['SECRET_KEY'] = os.getenv('SECRET_KEY', 'asdf#FGSgvasgf$5$WGT')

# Habilitar CORS para todas as rotas
CORS(app)

# Registrar blueprints
app.register_blueprint(user_bp, url_prefix='/api')
app.register_blueprint(firebase_bp, url_prefix='/api/firebase')
app.register_blueprint(contact_bp, url_prefix='/api')
app.register_blueprint(newsletter_bp, url_prefix='/api')

# Configuração do banco de dados
app.config['SQLALCHEMY_DATABASE_URI'] = f"sqlite:///{os.path.join(os.path.dirname(__file__), 'database', 'app.db')}"
app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
db.init_app(app)
with app.app_context():
    db.create_all()

# Rota de health check
@app.route('/health')
def health_check():
    return jsonify({
        'status': 'healthy',
        'service': 'Sentinela de Dados API',
        'version': '1.0.0'
    })

# Rota de informações da API
@app.route('/api')
def api_info():
    return jsonify({
        'message': 'Sentinela de Dados API',
        'version': '1.0.0',
        'endpoints': {
            'health': '/health',
            'firebase': '/api/firebase/*',
            'contact': '/api/contact',
            'newsletter': '/api/newsletter/*',
            'users': '/api/users/*'
        }
    })

# Handler para erros 404
@app.errorhandler(404)
def not_found(error):
    return jsonify({
        'success': False,
        'error': 'Endpoint não encontrado'
    }), 404

# Handler para erros 500
@app.errorhandler(500)
def internal_error(error):
    return jsonify({
        'success': False,
        'error': 'Erro interno do servidor'
    }), 500

@app.route('/', defaults={'path': ''})
@app.route('/<path:path>')
def serve(path):
    static_folder_path = app.static_folder
    if static_folder_path is None:
        return "Static folder not configured", 404

    if path != "" and os.path.exists(os.path.join(static_folder_path, path)):
        return send_from_directory(static_folder_path, path)
    else:
        index_path = os.path.join(static_folder_path, 'index.html')
        if os.path.exists(index_path):
            return send_from_directory(static_folder_path, 'index.html')
        else:
            return "index.html not found", 404

if __name__ == '__main__':
    app.run(host='0.0.0.0', port=5001, debug=True)

